A generic Menu/Config system
The menu struct takes in a HashMap of &str and the function that you
wanna call when that menu option is selected. The functions need to take
a mutable refrence and return a mutable refrence as the Menu expects to
perform CRUD operation, all of which requries mutability.

It can be used to create CRUD operations for any struct, Will also use
the same thing to modify extra config file.